Job_evaluation/requirements.txt
# Gmail and Google APIs
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# AI APIs
openai
anthropic

# Document Processing
python-docx
beautifulsoup4

# Data Processing
python-dotenv

# Utilities
requests